Add technical acceptance to feature request tmpl
This adds a new segment to the feature request template that allows the
engineering manager or lead developer to layout a set of criteria to
assist in providing direction for the implementation.
This will hopefully mirror the Acceptance Criteria field from the Client
segment, and allow things to be structured and communicated better in
the issues.